Updated static package, moved the model into a package called model Added schema to the gorp package
